# api/openapi.yaml
openapi: 3.0.3
info:
  title: Official Account Service API
  version: 1.0.0
  description: WeChat official-account authorization, article, material, publishing, callback, dashboard, and task-queue APIs.
servers:
  - url: http://localhost:8080
...
paths:
...
  /wechat/authorizer/{app_id}/callback:
    post:
      tags: [WeChatCallbacks]
      summary: Receive WeChat authorizer callbacks
      parameters: [{name: app_id, in: path, required: true, schema: {type: string}}, {$ref: "#/components/parameters/EncryptType"}, {$ref: "#/components/parameters/MsgSignature"}, {$ref: "#/components/parameters/Timestamp"}, {$ref: "#/components/parameters/Nonce"}]
      requestBody: {$ref: "#/components/requestBodies/XMLCallback"}
      responses: {"200": {$ref: "#/components/responses/CallbackOK"}, "400": {$ref: "#/components/responses/InvalidRequest"}}
...
  /api/v1/accounts/{id}/token-status:
    get:
      tags: [Accounts]
      summary: Get cached authorizer token status without token values
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}, {name: component_appid, in: query, required: true, schema: {type: string}}]
      responses: {"200": {description: Token status, content: {application/json: {schema: {$ref: "#/components/schemas/TokenStatus"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}, "404": {$ref: "#/components/responses/NotFound"}}
  /api/v1/articles:
    get:
      tags: [Articles]
      summary: List articles
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      responses: {"200": {$ref: "#/components/responses/ArticleList"}}
    post:
      tags: [Articles]
      summary: Create an article draft
      parameters: [{$ref: "#/components/parameters/TenantID"}]
      requestBody: {$ref: "#/components/requestBodies/CreateArticle"}
      responses: {"201": {description: Created article, content: {application/json: {schema: {$ref: "#/components/schemas/Article"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
  /api/v1/articles/{id}:
    get:
      tags: [Articles]
      summary: Get an article
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"200": {description: Article detail, content: {application/json: {schema: {$ref: "#/components/schemas/Article"}}}}, "404": {$ref: "#/components/responses/NotFound"}}
    put:
      tags: [Articles]
      summary: Update an article
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      requestBody: {$ref: "#/components/requestBodies/UpdateArticle"}
      responses: {"200": {description: Updated article, content: {application/json: {schema: {$ref: "#/components/schemas/Article"}}}}, "400": {$ref: "#/components/responses/InvalidRequest"}}
    delete:
      tags: [Articles]
      summary: Delete an article
      parameters: [{$ref: "#/components/parameters/TenantID"}, {$ref: "#/components/parameters/ID"}]
      responses: {"204": {description: Article deleted}, "404": {$ref: "#/components/responses/NotFound"}}
...
  /api/v1/task-queues/{queue}/archived-tasks:
    get:
      tags: [TaskQueues]
      summary: List archived asynq tasks with safe payload summaries
      parameters: [{$ref: "#/components/parameters/Queue"}, {name: limit, in: query, schema: {type: integer, minimum: 0, maximum: 100}}]
      responses: {"200": {$ref: "#/components/responses/ArchivedTaskList"}, "501": {$ref: "#/components/responses/NotImplemented"}}
  /api/v1/task-queues/{queue}/archived-tasks/{task_id}/retry:
    post:
      tags: [TaskQueues]
      summary: Move one archived task back to pending
      parameters: [{$ref: "#/components/parameters/Queue"}, {name: task_id, in: path, required: true, schema: {type: string}}]
      responses: {"200": {description: Task queued, content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: string, example: queued}}}}}}, "404": {$ref: "#/components/responses/NotFound"}}
components:
  parameters:
    TenantID: {name: X-Tenant-ID, in: header, required: true, schema: {type: string}}
    ID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    Queue: {name: queue, in: path, required: true, schema: {type: string, enum: [publish, token]}}
    EncryptType: {name: encrypt_type, in: query, schema: {type: string, enum: [aes]}}
    MsgSignature: {name: msg_signature, in: query, schema: {type: string}}
    Timestamp: {name: timestamp, in: query, schema: {type: string}}
    Nonce: {name: nonce, in: query, schema: {type: string}}
  requestBodies:
    XMLCallback: {required: true, content: {application/xml: {schema: {type: string}}}}
    CreateArticle: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateArticleRequest"}}}}
    UpdateArticle: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdateArticleRequest"}}}}
...
  responses:
...
    CallbackOK: {description: Callback accepted, content: {text/plain: {schema: {type: string, example: success}}}}
    InvalidRequest: {description: Invalid request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotImplemented: {description: Dependency not configured, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...
    ArticleList: {description: Article list, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: "#/components/schemas/Article"}}}}}}}
...
    ArchivedTaskList: {description: Archived task list, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: "#/components/schemas/ArchivedTask"}}}}}}}
  schemas:
...
    Error: {type: object, required: [error], properties: {error: {type: string, enum: [invalid_request, not_found, not_implemented, internal_error]}}}
...
    Article:
      type: object
      properties: {id: {type: integer, format: int64}, tenant_id: {type: string}, authorizer_id: {type: integer, format: int64}, title: {type: string}, author: {type: string}, digest: {type: string}, content_html: {type: string}, cover_media_asset_id: {type: integer, format: int64}, status: {type: string, enum: [draft, publishing, published, failed]}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}
    CreateArticleRequest:
      type: object
      required: [authorizer_id, title]
      properties: {authorizer_id: {type: integer, format: int64, minimum: 1}, title: {type: string}, author: {type: string}, digest: {type: string}, content_html: {type: string}}
    UpdateArticleRequest:
      allOf: [{$ref: "#/components/schemas/CreateArticleRequest"}, {type: object, properties: {cover_media_asset_id: {type: integer, format: int64, minimum: 0}}}]
...
    TokenStatus: {type: object, description: Does not expose access token or refresh token values., properties: {account_id: {type: integer, format: int64}, cache_hit: {type: boolean}, expires_at: {type: string, format: date-time}}}
...
    ArchivedTask:
      type: object
      description: Payload only includes safe summary fields and sensitive last_error values are redacted.
      properties: {id: {type: string}, queue: {type: string, enum: [publish, token]}, type: {type: string}, payload: {type: object, additionalProperties: {type: string}}, retried: {type: integer}, max_retry: {type: integer}, last_error: {type: string}, last_failed_at: {type: string, format: date-time}}